Move perception testing to offboard, separate 3d iou check
