+++ GEMstack/offboard/perception/sensorFusionTesting/pointpillars_ros/readme.md
@@ -1,0 +1,1 @@
+Please see GEMstack\offboard\perception\sensorFusionTesting\pointpillars_ros\scripts\pointpillars_ros_node.py for running tests.
