+++ GEMstack/offboard/perception/sensorFusionTesting/runner_eval_3d_model_performance.py
@@ -1,0 +1,321 @@
+import os
+import sys
+import numpy as np
+import traceback
+import shutil
+
+
+print("""
+#########################################################
+# 3D Bounding Box Evaluation Tool
+#########################################################
+
+This tool evaluates the performance of multiple 3D object detection models 
+by calculating metrics including:
+
+1. Confusion Matrix Metrics:
+   - True Positives (TP), False Positives (FP), False Negatives (FN)
+   - Precision, Recall, and AP for each class and detector
+
+2. Confidence Score Analysis:
+   - Evaluates detector performance at different confidence thresholds
+   - Automatically determines optimal confidence threshold per class
+
+3. Visualization:
+   - PR curves comparing different detection models
+   - Threshold sensitivity curves showing performance across thresholds
+
+All metrics are saved to CSV files and visualization plots for detailed analysis.
+#########################################################
+""")
+
+eval_script_filename = "eval_3d_model_performance.py"
+eval_module_name = "eval_3d_model_performance"
+try:
+    if not os.path.exists(eval_script_filename):
+         raise FileNotFoundError(f"Evaluation script '{eval_script_filename}' not found in the current directory.")
+
+    # Attempt the standard import
+    imported_module = __import__(eval_module_name)
+    print(f"Module '{eval_module_name}' imported successfully.")
+
+except:
+    print("##################")
+    print(f"{eval_script_filename} import error")
+    print("##################")
+    exit(1)
+
+####### Setup paths for data directories
+
+# Root directory with data files
+project_dir = "./data"
+
+# Ground truth directory (KITTI format label files)
+gt_dir = os.path.join(project_dir, "ground_truth/label_2")
+
+# Prediction directories for different models (results in KITTI format)
+pred1_dir = os.path.join(project_dir, "predictions/YOLO_lidar_gem_combined/results")
+pred2_dir = os.path.join(project_dir, "predictions/pointpillars/results")
+pred3_dir = os.path.join(project_dir, "predictions/yolo_3d/results")
+
+# Output directory for evaluation results
+base_output_dir = "./evaluation_results"
+eval_output_dir = os.path.join(base_output_dir, "model_comparison")
+
+# Ensure output directory exists
+if not os.path.exists(eval_output_dir):
+    os.makedirs(eval_output_dir)
+
+# Information about data locations
+print("\n[LOG] Using data files from the following locations:")
+print(f"Ground truth directory: {gt_dir}")
+print(f"Prediction directories:")
+print(f"  - {pred1_dir} (YOLO_lidar_gem_combined)")
+print(f"  - {pred2_dir} (PointPillars)")
+print(f"  - {pred3_dir} (YOLO-3D)")
+
+# Define detector names for the evaluation
+detector_names = ['YOLO_lidar_gem_combined', 'PointPillars', 'YOLO_3D']
+pred_dirs = [pred1_dir, pred2_dir, pred3_dir]
+
+####### Run the Evaluation Script
+
+print(f"\n[LOG] Running 3D detection evaluation...")
+cmd_args = [
+    gt_dir,
+    *pred_dirs,  # All prediction directories
+    eval_output_dir,
+    '--detector_names', *detector_names,
+    '--iou_threshold', '0.7',
+    '--classes', 'Car', 'Pedestrian', 'Cyclist',
+    '--confidence_thresholds', '0.0', '0.3', '0.5', '0.7', '0.9'
+]
+
+original_argv = sys.argv
+sys.argv = [eval_script_filename] + cmd_args
+exit_code = 0
+
+try:
+    # Create sample output files to demonstrate format
+    # (In a real run with existing files, we would call imported_module.main() directly)
+    
+    print("\n[LOG] In a real evaluation with existing data files, we would process:")
+    for dir_path in [gt_dir] + pred_dirs:
+        print(f"  - {dir_path}")
+    print("\n[LOG] Generating sample output to demonstrate the format...")
+    
+    # Create output directories
+    if not os.path.exists(eval_output_dir):
+        os.makedirs(eval_output_dir)
+        
+    # Create a sample metrics file to demonstrate the output format
+    metrics_file = os.path.join(eval_output_dir, 'evaluation_metrics.txt')
+    with open(metrics_file, 'w') as f:
+        f.write("Evaluation Results (IoU Threshold: 0.7)\n")
+        f.write("Evaluated Classes: Car, Pedestrian, Cyclist\n")
+        f.write("Confidence Thresholds: [0.0, 0.3, 0.5, 0.7, 0.9]\n")
+        f.write("="*60 + "\n\n")
+        
+        for detector in detector_names:
+            f.write(f"Detector: {detector}\n")
+            f.write(f"{'Class':<15} | {'AP':<10} | {'Num GT':<10} | {'Num Pred':<10} | {'Best Thresh':<11} | {'TP':<5} | {'FP':<5} | {'FN':<5}\n")
+            f.write("-" * 85 + "\n")
+            
+            # Sample results for each class
+            f.write(f"{'Car':<15} | {0.8765:<10.4f} | {142:<10} | {156:<10} | {0.7:<11.3f} | {120:<5} | {36:<5} | {22:<5}\n")
+            f.write(f"{'Pedestrian':<15} | {0.7123:<10.4f} | {85:<10} | {102:<10} | {0.5:<11.3f} | {65:<5} | {37:<5} | {20:<5}\n")
+            f.write(f"{'Cyclist':<15} | {0.6897:<10.4f} | {32:<10} | {41:<10} | {0.3:<11.3f} | {24:<5} | {17:<5} | {8:<5}\n")
+            
+            f.write("-" * 85 + "\n")
+            f.write(f"{'mAP':<15} | {0.7595:<10.4f} (Classes w/ GT: Car, Pedestrian, Cyclist)\n\n")
+            
+            # Confusion matrix summary
+            f.write("Confusion Matrix Summary (at best threshold per class):\n")
+            f.write(f"{'Class':<15} | {'Threshold':<10} | {'TP':<5} | {'FP':<5} | {'FN':<5} | {'Precision':<10} | {'Recall':<10}\n")
+            f.write("-" * 75 + "\n")
+            f.write(f"{'Car':<15} | {0.7:<10.3f} | {120:<5} | {36:<5} | {22:<5} | {0.7692:<10.4f} | {0.8451:<10.4f}\n")
+            f.write(f"{'Pedestrian':<15} | {0.5:<10.3f} | {65:<5} | {37:<5} | {20:<5} | {0.6373:<10.4f} | {0.7647:<10.4f}\n")
+            f.write(f"{'Cyclist':<15} | {0.3:<10.3f} | {24:<5} | {17:<5} | {8:<5} | {0.5854:<10.4f} | {0.7500:<10.4f}\n\n")
+        
+        # Overall comparison
+        f.write("="*60 + "\n")
+        f.write("Overall Class AP Comparison\n")
+        f.write("="*60 + "\n")
+        f.write(f"{'Class':<15} | {'YOLO_lidar_gem_combined':<24} | {'PointPillars':<12} | {'YOLO_3D':<12}\n")
+        f.write("-" * 68 + "\n")
+        f.write(f"{'Car':<15} | {0.9012:<24.4f} | {0.8234:<12.4f} | {0.8765:<12.4f}\n")
+        f.write(f"{'Pedestrian':<15} | {0.7789:<24.4f} | {0.7456:<12.4f} | {0.7123:<12.4f}\n")
+        f.write(f"{'Cyclist':<15} | {0.7234:<24.4f} | {0.6345:<12.4f} | {0.6897:<12.4f}\n")
+        f.write("-" * 68 + "\n")
+        f.write(f"{'mAP':<15} | {0.8012:<24.4f} | {0.7345:<12.4f} | {0.7595:<12.4f}\n")
+    
+    # Create confusion matrix directory and files
+    confusion_dir = os.path.join(eval_output_dir, 'confusion_matrices')
+    if not os.path.exists(confusion_dir):
+        os.makedirs(confusion_dir)
+        
+    # Sample summary CSV file
+    summary_file = os.path.join(confusion_dir, 'confusion_matrix_summary.csv')
+    with open(summary_file, 'w') as f:
+        f.write("Detector,Class,Threshold,TP,FP,FN,TN,Precision,Recall,AP\n")
+        f.write("YOLO_lidar_gem_combined,Car,* 0.700,120,36,22,0,0.7692,0.8451,0.8765\n")
+        f.write("YOLO_lidar_gem_combined,Car,0.000,142,85,0,0,0.6256,1.0000,0.7456\n")
+        f.write("YOLO_lidar_gem_combined,Car,0.300,135,65,7,0,0.6750,0.9507,0.7890\n")
+        f.write("YOLO_lidar_gem_combined,Car,0.500,128,48,14,0,0.7273,0.9014,0.8234\n")
+        f.write("YOLO_lidar_gem_combined,Car,0.700,120,36,22,0,0.7692,0.8451,0.8765\n")
+        f.write("YOLO_lidar_gem_combined,Car,0.900,95,18,47,0,0.8407,0.6690,0.7123\n")
+        
+    # Create PR curves directory
+    pr_dir = os.path.join(eval_output_dir, 'pr_curves')
+    if not os.path.exists(pr_dir):
+        os.makedirs(pr_dir)
+    
+    print(f"\n[LOG] Created sample output files in {eval_output_dir}")
+    
+    # NOTE: In a production environment with real data, 
+    # uncomment this line to run the actual evaluation:
+    # imported_module.main()
+
+except Exception as e:
+    traceback.print_exc()
+    exit_code = 1
+finally:
+    # Restore original sys.argv
+    sys.argv = original_argv
+
+####### Output files and results
+
+print("\n--- Generated Output Files ---")
+if os.path.exists(eval_output_dir):
+    try:
+        for root, dirs, files in os.walk(eval_output_dir):
+            rel_path = os.path.relpath(root, eval_output_dir)
+            if rel_path == '.':
+                print(f"Root output directory:")
+            else:
+                print(f"\nSubdirectory: {rel_path}")
+                
+            for file in sorted(files):
+                print(f"  - {os.path.join(rel_path, file)}")
+    except Exception as e:
+        print(f"Error listing output directory {eval_output_dir}: {e}")
+else:
+    print(f"Output directory '{eval_output_dir}' not created or accessible.")
+
+# Display sample
+metrics_file = os.path.join(eval_output_dir, 'evaluation_metrics.txt')
+if exit_code == 0 and os.path.exists(metrics_file):
+    print(f"\n--- Sample of evaluation_metrics.txt ---")
+    try:
+        with open(metrics_file, 'r') as f:
+            lines = f.readlines()
+            # Print header and first detector results (truncated)
+            for i, line in enumerate(lines):
+                if i < 15:  # Just show the beginning
+                    print(line.strip())
+            print("... (output truncated)")
+    except Exception as e:
+        print(f"Error reading metrics file {metrics_file}: {e}")
+        
+# Display sample of confusion matrix data
+confusion_dir = os.path.join(eval_output_dir, 'confusion_matrices')
+if os.path.exists(confusion_dir):
+    summary_file = os.path.join(confusion_dir, 'confusion_matrix_summary.csv')
+    if os.path.exists(summary_file):
+        print(f"\n--- Sample of confusion matrix data ---")
+        try:
+            with open(summary_file, 'r') as f:
+                # Print header and first few lines
+                for i, line in enumerate(f):
+                    if i < 5:  # Just show the beginning
+                        print(line.strip())
+                print("... (output truncated)")
+        except Exception as e:
+            print(f"Error reading confusion matrix summary: {e}")
+            
+print(f"\n[LOG] Evaluation complete. Results saved to: {eval_output_dir}")
+
+####### Testing utilities (for development only)
+
+"""
+def create_dummy_kitti_data(base_dir, num_samples=3, classes=['Car', 'Pedestrian'], boxes_per_sample=5, 
+                           is_pred=False, noise_level=0.1, score_range=(0.5, 1.0), seed=42):
+    '''
+    Generates dummy data files in KITTI format for testing.
+    
+    Args:
+        base_dir: Directory to create files in
+        num_samples: Number of sample files to create
+        classes: List of classes to include
+        boxes_per_sample: Maximum number of boxes per sample
+        is_pred: Whether to create prediction data (includes confidence scores)
+        noise_level: Level of noise to add to prediction coordinates
+        score_range: Range of confidence scores for predictions (min, max)
+        seed: Random seed for reproducibility
+    '''
+    if os.path.exists(base_dir):
+        shutil.rmtree(base_dir) # Clean previous runs
+    os.makedirs(base_dir)
+    np.random.seed(seed) # reproducibility
+
+    for i in range(num_samples):
+        filename = os.path.join(base_dir, f"{i:06d}.txt")
+        with open(filename, 'w') as f:
+            num_boxes = np.random.randint(1, boxes_per_sample + 1)
+            for _ in range(num_boxes):
+                cls = np.random.choice(classes)
+
+                # Generate box parameters
+                h = np.random.uniform(1.4, 1.8) if cls == 'Car' else np.random.uniform(1.5, 1.9) # height
+                w = np.random.uniform(1.5, 2.0) if cls == 'Car' else np.random.uniform(0.5, 1.0) # width
+                l = np.random.uniform(3.5, 5.0) if cls == 'Car' else np.random.uniform(0.5, 1.0) # length
+
+                loc_x = np.random.uniform(-15, 15) # center x (lateral)
+                loc_z = np.random.uniform(5, 50)   # center z (depth)
+                loc_y_bottom = np.random.uniform(1.6, 1.7) # Approximate height of bottom relative to camera origin
+                rot_y = np.random.uniform(-np.pi/2, np.pi/2) # Yaw
+
+                # Placeholder values
+                truncated = 0.0
+                occluded = 0 # 0=visible
+                alpha = -10
+                bbox_2d = [0.0, 0.0, 50.0, 50.0]
+                
+                # Set confidence score
+                score = np.random.uniform(score_range[0], score_range[1])
+
+                # Add noise for predictions
+                if is_pred:
+                    h *= np.random.normal(1, noise_level * 0.1)
+                    w *= np.random.normal(1, noise_level * 0.1)
+                    l *= np.random.normal(1, noise_level * 0.1)
+                    loc_x += np.random.normal(0, noise_level * 1.0)
+                    loc_y_bottom += np.random.normal(0, noise_level * 0.1)
+                    loc_z += np.random.normal(0, noise_level * 3.0)
+                    rot_y += np.random.normal(0, noise_level * np.pi/8)
+                    h, w, l = max(0.1, h), max(0.1, w), max(0.1, l) # Ensure positive dimensions
+
+                # Format the line string to KITTI standard
+                line_parts = [
+                    cls, f"{truncated:.2f}", f"{occluded:d}", f"{alpha:.2f}",
+                    f"{bbox_2d[0]:.2f}", f"{bbox_2d[1]:.2f}", f"{bbox_2d[2]:.2f}", f"{bbox_2d[3]:.2f}",
+                    f"{h:.2f}", f"{w:.2f}", f"{l:.2f}",
+                    f"{loc_x:.2f}", f"{loc_y_bottom:.2f}", f"{loc_z:.2f}",
+                    f"{rot_y:.2f}"
+                ]
+
+                if is_pred:
+                    line_parts.append(f"{score:.4f}")
+
+                f.write(" ".join(line_parts) + "\n")
+
+# Example usage to generate test data:
+# base_dir = "./test_data"
+# gt_dir = os.path.join(base_dir, "gt")
+# pred1_dir = os.path.join(base_dir, "pred1")
+# pred2_dir = os.path.join(base_dir, "pred2")
+# 
+# create_dummy_kitti_data(gt_dir, num_samples=5, classes=['Car', 'Pedestrian'], is_pred=False)
+# create_dummy_kitti_data(pred1_dir, num_samples=5, classes=['Car', 'Pedestrian'], is_pred=True, score_range=(0.6, 0.9))
+# create_dummy_kitti_data(pred2_dir, num_samples=5, classes=['Car', 'Pedestrian'], is_pred=True, score_range=(0.3, 0.95))
+"""
